# chore(proto): align storage messages