--- a/generated/storage/administration_event.pb.go
+++ b/generated/storage/administration_event.pb.go
@@ -135,22 +135,23 @@ func (AdministrationEventLevel) EnumDescriptor() ([]byte, []int) {
 //
 // Refer to v1.AdministrationEvent for a more detailed doc.
 type AdministrationEvent struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
+	state          protoimpl.MessageState
+	Resource       *AdministrationEvent_Resource `protobuf:"bytes,7,opt,name=resource,proto3" json:"resource,omitempty"`
+	LastOccurredAt *timestamppb.Timestamp        `protobuf:"bytes,9,opt,name=last_occurred_at,json=lastOccurredAt,proto3" json:"last_occurred_at,omitempty" search:"Last Updated,hidden"` // @gotags: search:"Last Updated,hidden"
+	CreatedAt      *timestamppb.Timestamp        `protobuf:"bytes,10,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty" search:"Created Time,hidden"`                 // @gotags: search:"Created Time,hidden"
 
 	// The id is a UUIDv5 generated deterministically from the tuple (type, level, domain,
 	// resource_type, resource_id, message). It is used for deduplication of events.
-	Id             string                        `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty" sql:"pk,type(uuid)"`                                              // @gotags: sql:"pk,type(uuid)"
-	Type           AdministrationEventType       `protobuf:"varint,2,opt,name=type,proto3,enum=storage.AdministrationEventType" json:"type,omitempty" search:"Event Type,hidden"`    // @gotags: search:"Event Type,hidden"
-	Level          AdministrationEventLevel      `protobuf:"varint,3,opt,name=level,proto3,enum=storage.AdministrationEventLevel" json:"level,omitempty" search:"Event Level,hidden"` // @gotags: search:"Event Level,hidden"
-	Message        string                        `protobuf:"bytes,4,opt,name=message,proto3" json:"message,omitempty"`
-	Hint           string                        `protobuf:"bytes,5,opt,name=hint,proto3" json:"hint,omitempty"`
-	Domain         string                        `protobuf:"bytes,6,opt,name=domain,proto3" json:"domain,omitempty" search:"Event Domain,hidden"` // @gotags: search:"Event Domain,hidden"
-	Resource       *AdministrationEvent_Resource `protobuf:"bytes,7,opt,name=resource,proto3" json:"resource,omitempty"`
-	NumOccurrences int64                         `protobuf:"varint,8,opt,name=num_occurrences,json=numOccurrences,proto3" json:"num_occurrences,omitempty" search:"Event Occurrence,hidden"`  // @gotags: search:"Event Occurrence,hidden"
-	LastOccurredAt *timestamppb.Timestamp        `protobuf:"bytes,9,opt,name=last_occurred_at,json=lastOccurredAt,proto3" json:"last_occurred_at,omitempty" search:"Last Updated,hidden"` // @gotags: search:"Last Updated,hidden"
-	CreatedAt      *timestamppb.Timestamp        `protobuf:"bytes,10,opt,name=created_at,json=createdAt,proto3" json:"created_at,omitempty" search:"Created Time,hidden"`                 // @gotags: search:"Created Time,hidden"
+	Id            string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty" sql:"pk,type(uuid)"` // @gotags: sql:"pk,type(uuid)"
+	Message       string `protobuf:"bytes,4,opt,name=message,proto3" json:"message,omitempty"`
+	Hint          string `protobuf:"bytes,5,opt,name=hint,proto3" json:"hint,omitempty"`
+	Domain        string `protobuf:"bytes,6,opt,name=domain,proto3" json:"domain,omitempty" search:"Event Domain,hidden"` // @gotags: search:"Event Domain,hidden"
+	unknownFields protoimpl.UnknownFields
+
+	NumOccurrences int64 `protobuf:"varint,8,opt,name=num_occurrences,json=numOccurrences,proto3" json:"num_occurrences,omitempty" search:"Event Occurrence,hidden"` // @gotags: search:"Event Occurrence,hidden"
+	sizeCache      protoimpl.SizeCache
+	Type           AdministrationEventType  `protobuf:"varint,2,opt,name=type,proto3,enum=storage.AdministrationEventType" json:"type,omitempty" search:"Event Type,hidden"`     // @gotags: search:"Event Type,hidden"
+	Level          AdministrationEventLevel `protobuf:"varint,3,opt,name=level,proto3,enum=storage.AdministrationEventLevel" json:"level,omitempty" search:"Event Level,hidden"` // @gotags: search:"Event Level,hidden"
 }
 
 func (x *AdministrationEvent) Reset() {
@@ -256,13 +257,14 @@ func (x *AdministrationEvent) GetCreatedAt() *timestamppb.Timestamp {
 }
 
 type AdministrationEvent_Resource struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state protoimpl.MessageState
+
+	Type          string `protobuf:"bytes,1,opt,name=type,proto3" json:"type,omitempty" search:"Resource Type,hidden"` // @gotags: search:"Resource Type,hidden"
+	Id            string `protobuf:"bytes,2,opt,name=id,proto3" json:"id,omitempty"`
+	Name          string `protobuf:"bytes,3,opt,name=name,proto3" json:"name,omitempty"`
 	unknownFields protoimpl.UnknownFields
 
-	Type string `protobuf:"bytes,1,opt,name=type,proto3" json:"type,omitempty" search:"Resource Type,hidden"` // @gotags: search:"Resource Type,hidden"
-	Id   string `protobuf:"bytes,2,opt,name=id,proto3" json:"id,omitempty"`
-	Name string `protobuf:"bytes,3,opt,name=name,proto3" json:"name,omitempty"`
+	sizeCache protoimpl.SizeCache
 }
 
 func (x *AdministrationEvent_Resource) Reset() {