… generated/storage/alert.pb.go
… generated/storage/cloud_source.pb.go
… generated/storage/cluster.pb.go
… generated/storage/compliance.pb.go
… generated/storage/compliance_integration.pb.go
… generated/storage/compliance_operator.pb.go
… generated/storage/compliance_operator_v2.pb.go
… generated/storage/config.pb.go
… generated/storage/container_runtime.pb.go
… generated/storage/cve.pb.go
… generated/storage/declarative_config_health.pb.go
… generated/storage/delegated_registry_config.pb.go
… generated/storage/deployment.pb.go
… generated/storage/discovered_cluster.pb.go
… generated/storage/external_backup.pb.go
… generated/storage/group.pb.go
… generated/storage/hash.pb.go
… generated/storage/http_endpoint.pb.go
… generated/storage/image.pb.go
… generated/storage/image_component.pb.go
… generated/storage/image_integration.pb.go
… generated/storage/installation.pb.go
… generated/storage/integration_health.pb.go
… generated/storage/kube_event.pb.go
… generated/storage/labels.pb.go
… generated/storage/log.pb.go
… generated/storage/mitre.pb.go
… generated/storage/namespace_metadata.pb.go
… generated/storage/network_baseline.pb.go
… generated/storage/network_flow.pb.go