### Description

Aligning structs can save ~1% of memory for alerts. Maybe it's worth to consider to enable it.

```
 benchstat -table "PGX"  -confidence 0.030  old.txt new.txt
PGX: 
                       │   old.txt   │              new.txt               │
                       │   sec/op    │   sec/op     vs base               │
Many/get_1_alerts-8      161.2µ ± 0%   161.0µ ± 0%       ~ (p=0.684 n=10)
Many/get_2_alerts-8      177.1µ ± 0%   180.1µ ± 1%       ~ (p=0.796 n=10)
Many/get_4_alerts-8      217.5µ ± 0%   218.8µ ± 0%       ~ (p=0.684 n=10)
Many/get_8_alerts-8      261.8µ ± 0%   265.9µ ± 0%       ~ (p=0.063 n=10)
Many/get_16_alerts-8     358.3µ ± 0%   360.5µ ± 0%       ~ (p=0.481 n=10)
Many/get_32_alerts-8     560.2µ ± 1%   557.0µ ± 1%       ~ (p=0.579 n=10)
Many/get_64_alerts-8     943.1µ ± 1%   927.2µ ± 0%       ~ (p=0.684 n=10)
Many/get_128_alerts-8    1.649m ± 0%   1.653m ± 1%       ~ (p=0.912 n=10)
Many/get_256_alerts-8    3.042m ± 0%   2.979m ± 0%       ~ (p=0.393 n=10)
Many/get_512_alerts-8    5.717m ± 1%   5.737m ± 0%       ~ (p=0.529 n=10)
Many/get_1024_alerts-8   11.12m ± 0%   10.89m ± 0%       ~ (p=0.143 n=10)
Many/get_2048_alerts-8   22.48m ± 0%   22.02m ± 0%       ~ (p=0.105 n=10)
Many/get_4096_alerts-8   43.82m ± 0%   42.77m ± 0%       ~ (p=0.123 n=10)
Many/get_8192_alerts-8   87.94m ± 0%   86.13m ± 0%       ~ (p=0.218 n=10)
geomean                  1.908m        1.897m       -0.59%

                       │   old.txt    │               new.txt               │
                       │     B/op     │     B/op      vs base               │
Many/get_1_alerts-8      9.685Ki ± 0%   9.606Ki ± 0%  -0.81% (p=0.000 n=10)
Many/get_2_alerts-8      15.05Ki ± 0%   14.90Ki ± 0%  -1.03% (p=0.000 n=10)
Many/get_4_alerts-8      25.80Ki ± 0%   25.49Ki ± 0%  -1.21% (p=0.000 n=10)
Many/get_8_alerts-8      48.06Ki ± 0%   47.43Ki ± 0%  -1.30% (p=0.000 n=10)
Many/get_16_alerts-8     92.86Ki ± 0%   91.61Ki ± 0%  -1.35% (p=0.000 n=10)
Many/get_32_alerts-8     183.7Ki ± 0%   181.1Ki ± 0%  -1.37% (p=0.000 n=10)
Many/get_64_alerts-8     364.3Ki ± 0%   359.2Ki ± 0%  -1.38% (p=0.000 n=10)
Many/get_128_alerts-8    724.7Ki ± 0%   714.6Ki ± 0%  -1.39% (p=0.000 n=10)
Many/get_256_alerts-8    1.421Mi ± 0%   1.401Mi ± 0%  -1.37% (p=0.000 n=10)
Many/get_512_alerts-8    2.834Mi ± 0%   2.794Mi ± 0%  -1.38% (p=0.000 n=10)
Many/get_1024_alerts-8   5.694Mi ± 0%   5.618Mi ± 0%  -1.34% (p=0.000 n=10)
Many/get_2048_alerts-8   11.70Mi ± 0%   11.54Mi ± 0%  -1.34% (p=0.000 n=10)
Many/get_4096_alerts-8   23.28Mi ± 0%   22.96Mi ± 0%  -1.34% (p=0.000 n=10)
Many/get_8192_alerts-8   46.77Mi ± 0%   46.14Mi ± 0%  -1.34% (p=0.000 n=10)
geomean                  556.6Ki        549.5Ki       -1.28%

                       │   old.txt   │               new.txt                │
                       │  allocs/op  │  allocs/op   vs base                 │
Many/get_1_alerts-8       136.0 ± 0%    136.0 ± 0%       ~ (p=1.000 n=10) ¹
Many/get_2_alerts-8       187.0 ± 0%    187.0 ± 0%       ~ (p=1.000 n=10) ¹
Many/get_4_alerts-8       288.0 ± 0%    288.0 ± 0%       ~ (p=1.000 n=10) ¹
Many/get_8_alerts-8       491.0 ± 0%    491.0 ± 0%       ~ (p=1.000 n=10) ¹
Many/get_16_alerts-8      894.0 ± 0%    894.0 ± 0%       ~ (p=1.000 n=10) ¹
Many/get_32_alerts-8     1.698k ± 0%   1.698k ± 0%       ~ (p=1.000 n=10) ¹
Many/get_64_alerts-8     3.302k ± 0%   3.302k ± 0%       ~ (p=1.000 n=10) ¹
Many/get_128_alerts-8    6.505k ± 0%   6.505k ± 0%       ~ (p=1.000 n=10) ¹
Many/get_256_alerts-8    12.91k ± 0%   12.91k ± 0%       ~ (p=1.000 n=10)
Many/get_512_alerts-8    25.71k ± 0%   25.71k ± 0%       ~ (p=1.000 n=10)
Many/get_1024_alerts-8   51.32k ± 0%   51.32k ± 0%       ~ (p=0.625 n=10)
Many/get_2048_alerts-8   102.5k ± 0%   102.5k ± 0%       ~ (p=0.474 n=10)
Many/get_4096_alerts-8   205.0k ± 0%   205.0k ± 0%       ~ (p=0.350 n=10)
Many/get_8192_alerts-8   409.8k ± 0%   409.8k ± 0%       ~ (p=1.000 n=10)
geomean                  5.384k        5.384k       +0.00%
```

---
- [x] CHANGELOG update is not needed
- [x] Documentation is not needed
### Testing
- [x] inspected CI results
#### Automated testing
- [x] modified existing tests
- [x] contributed **no automated tests**
#### How I validated my change
CI